+++ cgame/cg_lua.h
@@ -48,7 +48,7 @@ typedef struct jplua_plugin_s
 	char name[32];
 	char version[8];
 	unsigned int requiredJPLuaVersion;
-	unsigned int UID;
+	intptr_t UID;
 	int handle;
 
 	int eventListeners[JPLUA_EVENT_MAX]; // references to listener functions in lua stored in the registry
